게시판 보완 및 이메일 인증 추가

+++ src/main/java/com/example/knockknock/domain/post/service/PostService.java
@@ -1,6 +1,8 @@
 package com.example.knockknock.domain.post.service;
 
 import com.example.knockknock.domain.hashtag.dto.HashtagRegisterRequestDto;
+import com.example.knockknock.domain.member.security.UserDetailsImpl;
+import com.example.knockknock.domain.member.service.MemberIsLoginService;
 import com.example.knockknock.domain.post.dto.request.*;
 import com.example.knockknock.domain.post.dto.response.*;
 import com.example.knockknock.domain.hashtag.entity.Hashtag;
@@ -22,19 +24,19 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
-    private final MemberRepository memberRepository;
+    private final MemberIsLoginService memberIsLoginService;
 
     private final S3Service s3Service;
 
     @Transactional
-    public void createPost(PostCreateRequestDto request, List<MultipartFile> images) {
-        Member member =  memberRepository.findById(request.getMemberId())
-                .orElseThrow(() -> new GlobalException(GlobalErrorCode.MEMBER_NOT_FOUND));
+    public void createPost(PostCreateRequestDto request, List<MultipartFile> images, UserDetailsImpl userDetails) {
+        Member member = memberIsLoginService.isLogin(userDetails);
         Post post = Post.builder()
                 .member(member)
                 .boardType(request.getBoardType())
@@ -66,9 +68,29 @@ public class PostService {
     }
 
     @Transactional
-    public void updatePost(Long postId , PostUpdateRequestDto request, List<MultipartFile> images) {
+    public List<MyPostsResponseDto> getMyPosts(UserDetailsImpl userDetails){
+        Member member = memberIsLoginService.isLogin(userDetails);
+        List<Post> myPosts = postRepository.findByMember(member);
+        List<MyPostsResponseDto> myPostsResponseDtos = myPosts.stream()
+                .map(MyPostsResponseDto::of)
+                .collect(Collectors.toList());
+
+        return myPostsResponseDtos;
+    }
+
+    @Transactional
+    public void updatePost(Long postId, PostUpdateRequestDto request, List<MultipartFile> images, UserDetailsImpl userDetails) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new GlobalException(GlobalErrorCode.POST_NOT_FOUND));
+        // 현재 로그인한 멤버 정보 가져오기
+        Long currentMemberId = userDetails.getUser().getMemberId();
+
+        // 게시글 작성자 정보 가져오기
+        Long postAuthorId = post.getMember().getMemberId();
+
+        if (!currentMemberId.equals(postAuthorId)) {
+            throw new GlobalException(GlobalErrorCode.PERMISSION_DENIED);
+        }
         post.updatePost(request);
 
         //이미지 추가
@@ -93,9 +115,19 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, UserDetailsImpl userDetails) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new GlobalException(GlobalErrorCode.POST_NOT_FOUND));
+
+        // 현재 로그인한 멤버 정보 가져오기
+        Long currentMemberId = userDetails.getUser().getMemberId();
+
+        // 게시글 작성자 정보 가져오기
+        Long postAuthorId = post.getMember().getMemberId();
+
+        if (!currentMemberId.equals(postAuthorId)) {
+            throw new GlobalException(GlobalErrorCode.PERMISSION_DENIED);
+        }
         postRepository.delete(post);
     }
